requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.7
langchain-google-genai==0.0.9
pymysql==1.1.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
python-dotenv==1.0.1
pydantic==2.6.1 